pip_requirements.txt
Django==1.6.5
celery==3.1.13
ipython==2.1.0
lxml==3.3.5
mongoengine==0.8.7
pymongo==2.7.1
python-dateutil==2.4.1
pytz==2014.3
requests==2.3.0
psycopg2==2.5.3
Pillow==2.4.0
pdfkit==0.4.1
django-grappelli==2.5.3
jsonfield==0.9.22
South==1.0
django-filter==0.9.1
djangorestframework==3.0.0
django-wkhtmltopdf==2.0.2
pycrypto==2.6.1
django-oauth-toolkit==0.8.1
django-cors-headers==1.1.0
